move config set up to test file so that it doesnt seal it when starting the chain
