+++ app/setup_test.go
@@ -1,0 +1,17 @@
+package app
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	cmdcfg "github.com/evmos/ethermint/cmd/config"
+)
+
+func init() {
+	SetupConfig()
+}
+func SetupConfig() {
+	// set the address prefixes
+	config := sdk.GetConfig()
+	cmdcfg.SetBech32Prefixes(config)
+	cmdcfg.SetBip44CoinType(config)
+	config.Seal()
+}
